requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic[email]==2.8.2
asyncmy>=0.2,<1.0
